+++ src/wp/ap/xp/ap_Menu_Functions.h
@@ -83,6 +83,7 @@ Defun_EV_GetMenuItemState_Fn(ap_GetState_InFootnote);
 Defun_EV_GetMenuItemState_Fn(ap_GetState_InImage);
 Defun_EV_GetMenuItemState_Fn(ap_GetState_SetPosImage);
 Defun_EV_GetMenuItemState_Fn(ap_GetState_InFrame);
+Defun_EV_GetMenuItemState_Fn(ap_GetState_ObjSelected);
 Defun_EV_GetMenuItemState_Fn(ap_GetState_AlwaysDisabled); // REMOVE ME
 Defun_EV_GetMenuItemState_Fn(ap_GetState_Recent);
 Defun_EV_GetMenuItemState_Fn(ap_GetState_Zoom);